More happy about nested type schema
