+++ tests/AvroSourceGenerator.Tests.Apache/LogicalTests.Verify_Decimal_Fixed#00.verified.txt
@@ -19,7 +19,9 @@ namespace SchemaNamespace
           "type": "fixed",
           "name": "Decimal",
           "namespace": "SchemaNamespace",
-          "size": 20
+          "size": 20,
+          "precision": 4,
+          "scale": 2
         }
         """);
     }
