+++ tests/AvroSourceGenerator.Tests/NestedType.Works.verified.txt
@@ -1,0 +1,11 @@
+﻿// <auto-generated/>
+#pragma warning disable CS8618, CS8633, CS8714, CS8775, CS8981
+#nullable enable
+    [global::System.CodeDom.Compiler.GeneratedCode("AvroSourceGenerator", "1.0.0.0")]
+    public partial record SomeSchema
+    {
+        public required string displayName { get; init; }
+    }
+    
+#nullable restore
+#pragma warning restore CS8618, CS8633, CS8714, CS8775
